# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Releasy API",
    "description": "Releasy API surface for releases, artifacts, downloads, and admin ops.",
    "license": {
      "name": ""
    },
    "version": "0.1.3"
  },
  "paths": {
...
    "/v1/admin/customers": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List customers",
        "description": "Lists customers with optional filters. Requires platform_support or platform_admin role.",
        "operationId": "list_customers",
        "parameters": [
          {
            "name": "customer_id",
            "in": "query",
            "description": "Optional exact customer id filter",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "Optional name filter (case-insensitive contains)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "plan",
            "in": "query",
            "description": "Optional plan filter (case-insensitive exact match)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size (default 50, max 200)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Page offset (default 0)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customers list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCustomerListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Admin auth not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_key": []
          },
          {
            "operator_jwt": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create a customer",
        "description": "Creates a new customer record. Requires platform_admin role.",
        "operationId": "admin_create_customer",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Idempotency key for safe retries.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCreateCustomerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Customer created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCreateCustomerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Admin auth not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_key": []
          },
          {
            "operator_jwt": []
          }
        ]
      }
    },
    "/v1/admin/customers/{customer_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get a customer",
        "description": "Fetches a customer record. Requires platform_support or platform_admin role.",
        "operationId": "get_customer",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "description": "Customer identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCustomerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Admin auth not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_key": []
          },
          {
            "operator_jwt": []
          }
        ]
      },
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Update a customer",
        "description": "Updates customer name, plan, or suspension. Requires platform_admin role.",
        "operationId": "update_customer",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "description": "Customer identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateCustomerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Customer updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCustomerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Admin auth not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_key": []
          },
          {
            "operator_jwt": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminCreateCustomerRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "plan": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "AdminCreateCustomerResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "plan": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "AdminCustomerListResponse": {
        "type": "object",
        "required": [
          "customers",
          "limit",
          "offset"
        ],
        "properties": {
          "customers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminCustomerResponse"
            }
          },
          "limit": {
            "type": "integer",
            "format": "int64"
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "AdminCustomerResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "plan": {
            "type": [
              "string",
              "null"
            ]
          },
          "suspended_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
...
      "AdminUpdateCustomerRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "plan": {
            "type": [
              "string",
              "null"
            ]
          },
          "suspended": {
            "type": [
              "boolean",
              "null"
            ]
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "admin_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-releasy-admin-key"
      },
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-releasy-api-key"
      },
      "operator_jwt": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}